Minor refactors
Move default implementation classes to "impl"
Move default implementation instance fields into interfaces (DefaultXXX.INSTANCE -> XXX.DEFAULT)
