+++ fabric-renderer-registries-v1/src/main/java/net/fabricmc/fabric/api/client/rendereregistry/v1/item/DurabilityBarProperties.java
@@ -18,7 +18,11 @@ package net.fabricmc.fabric.api.client.rendereregistry.v1.item;
 
 import net.minecraft.item.ItemStack;
 
+import net.fabricmc.fabric.impl.client.renderer.registry.DefaultDurabilityBarProperties;
+
 public interface DurabilityBarProperties {
+	DurabilityBarProperties DEFAULT = new DefaultDurabilityBarProperties();
+
 	int getCount(ItemStack stack);
 	boolean isVisible(ItemStack stack, int index);
 	float getFillFactor(ItemStack stack, int index);
